+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.1
 
 - Fixed a crash when opening the Stems tab in Preferences on installed builds.
+- Fixed first-use stem separation to download the RoFormer quality models (vocals + drums/bass) together by default, instead of the backup model.
 
 ## 1.0.0